+++ src/main/java/org/bboxdb/performance/BenchmarkInsertPerformance.java
@@ -61,9 +61,9 @@ public class BenchmarkInsertPerformance extends AbstractBenchmark {
 	
 		// Insert the tuples
 		for(; insertedTuples.get() < tuples; insertedTuples.incrementAndGet()) {
-			final float x = (float) Math.abs(bbBoxRandom.nextFloat() % 100000.0 * 1000);
-			final float y = (float) Math.abs(bbBoxRandom.nextFloat() % 100000.0 * 1000);
-			final float z = (float) Math.abs(bbBoxRandom.nextFloat() % 100000.0 * 1000);
+			final double x = Math.abs(bbBoxRandom.nextFloat() % 100000.0 * 1000);
+			final double y = Math.abs(bbBoxRandom.nextFloat() % 100000.0 * 1000);
+			final double z = Math.abs(bbBoxRandom.nextFloat() % 100000.0 * 1000);
 			
 			final BoundingBox boundingBox = new BoundingBox(x, x+1, y, y+1, z, z+1);
